+++ manual_test.sh
@@ -8,6 +8,10 @@ echo "== TESTING -p tell a great joke -t 0 ==" && \
 cargo run -- -p tell a great joke -t 0 && \
 echo "== TESTING -m gpt-3.5-turbo-0301 ==" && \
 cargo run -- -m gpt-3.5-turbo-0301 -p language is elixir create a function that adds two numbers && \
+echo "== TESTING -c ==" && \
+cargo run -- -c && \
+echo "== TESTING -c Roleplay as Shakespeare ==" && \
+cargo run -- -c Roleplay as Shakespeare && \
 echo ""
 echo "== ======= =="
 echo "== SUCCESS =="