+++ src/main.rs
@@ -12,7 +12,7 @@ mod request;
 const APP: &str = "GPTO";
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 const AUTHOR: &str = "Alan Vardy <alan@vardy.cc>";
-const ABOUT: &str = "A tiny unofficial OpenAI GPT3 client";
+const ABOUT: &str = "A tiny unofficial OpenAI client";
 
 pub const MODEL_DEFAULT: &str = "gpt-3.5-turbo";
 pub const MODEL_HELP: &str = "
@@ -41,12 +41,15 @@ pub const SUFFIX_HELP: &str =
 pub const CONFIG_HELP: &str =
     "Absolute path of configuration. Defaults to $XDG_CONFIG_HOME/gpto.cfg";
 pub const PROMPT_HELP: &str = "The prompt(s) to generate completions for";
+pub const CONVERSATION_HELP: &str =
+    "Start a conversation with an optional description of the bot's role";
 pub const NUMBER_HELP: &str = "How many completions to generate for each prompt. Defaults to 1";
 pub const ECHO_DEFAULT: bool = false;
 pub const MODELS_HELP: &str = "Returns a list of models from OpenAI";
 
 pub struct Arguments<'a> {
     prompt: Option<String>,
+    conversation: Option<String>,
     suffix: Option<String>,
     number: Option<u8>,
     temperature: Option<f32>,
@@ -61,8 +64,19 @@ fn main() {
         .author(AUTHOR)
         .about(ABOUT);
     let matches = app
-        .arg(flag_string("prompt", 'p', PROMPT_HELP))
-        .arg(flag_string("suffix", 's', SUFFIX_HELP))
+        .arg(flag_string("prompt", 'p', "Prompt text", PROMPT_HELP))
+        .arg(flag_nullable_string(
+            "conversation",
+            'c',
+            "Instructions to bot",
+            CONVERSATION_HELP,
+        ))
+        .arg(flag_string(
+            "suffix",
+            's',
+            "Text to be appended to end of response",
+            SUFFIX_HELP,
+        ))
         .arg(flag_float("temperature", 't', TEMPERATURE_HELP))
         .arg(flag_integer("number", 'n', NUMBER_HELP))
         .arg(flag_float("top_p", 'k', TOP_P_HELP))
@@ -83,6 +97,7 @@ fn main() {
     let arguments = Arguments {
         prompt: join_string(matches.clone(), "prompt"),
         suffix: join_string(matches.clone(), "suffix"),
+        conversation: join_string(matches.clone(), "conversation"),
         config_path: matches.get_one::<String>("config").map(|s| s.as_str()),
         model: matches.get_one::<String>("model").map(|s| s.as_str()),
         number: matches.get_one::<u8>("number").map(|s| s.to_owned()),
@@ -107,8 +122,15 @@ fn dispatch(arguments: Arguments) -> Result<String, String> {
 
     match arguments {
         Arguments {
-            prompt: Some(_), ..
+            prompt: Some(_),
+            conversation: None,
+            ..
         } => request::completions(arguments, config),
+        Arguments {
+            conversation: Some(_),
+            prompt: None,
+            ..
+        } => request::conversation(arguments, config),
         Arguments {
             prompt: None,
             config_path: _,
@@ -117,6 +139,7 @@ fn dispatch(arguments: Arguments) -> Result<String, String> {
             suffix: None,
             temperature: None,
             top_p: None,
+            conversation: None,
         } => Err(String::from(
             "gtpo cannot be run without parameters. To see available parameters use --help",
         )),
@@ -141,7 +164,7 @@ fn flag_string_no_spaces(
         .help(help)
 }
 
-fn flag_string(long: &'static str, short: char, help: &'static str) -> Arg {
+fn flag_string(long: &'static str, short: char, value: &'static str, help: &'static str) -> Arg {
     Arg::new(long)
         .short(short)
         .long(long)
@@ -149,6 +172,24 @@ fn flag_string(long: &'static str, short: char, help: &'static str) -> Arg {
         .action(ArgAction::Append)
         .num_args(1..)
         .value_parser(clap::value_parser!(String))
+        .value_name(value)
+        .help(help)
+}
+
+fn flag_nullable_string(
+    long: &'static str,
+    short: char,
+    value: &'static str,
+    help: &'static str,
+) -> Arg {
+    Arg::new(long)
+        .short(short)
+        .long(long)
+        .required(false)
+        .action(ArgAction::Append)
+        .num_args(0..)
+        .value_parser(clap::value_parser!(String))
+        .value_name(value)
         .help(help)
 }
 fn flag_float(long: &'static str, short: char, help: &'static str) -> Arg {